--- test/test_init/main.cpp
@@ -1,34 +1,0 @@
-#include <Arduino.h>
-#include <unity.h>
-
-void testInitialisation(void);
-void testFuelScheduleInit(void);
-void testIgnitionScheduleInit(void);
-
-#define UNITY_EXCLUDE_DETAILS
-
-void setup()
-{
-    pinMode(LED_BUILTIN, OUTPUT);
-
-    // NOTE!!! Wait for >2 secs
-    // if board doesn't support software reset via Serial.DTR/RTS
-    delay(2000);
-
-    UNITY_BEGIN();    // IMPORTANT LINE!
-
-    testInitialisation();
-    testFuelScheduleInit();
-    testIgnitionScheduleInit();
-
-    UNITY_END(); // stop unit testing
-}
-
-void loop()
-{
-    // Blink to indicate end of test
-    digitalWrite(LED_BUILTIN, HIGH);
-    delay(250);
-    digitalWrite(LED_BUILTIN, LOW);
-    delay(250);
-}
