+++ test/test_misc/tests_misc.cpp
@@ -1,6 +1,7 @@
 #include <Arduino.h>
 #include <unity.h>
 
+#include "tests_init.h"
 #include "tests_tables.h"
 #include "test_table2d.h"
 
@@ -16,6 +17,7 @@ void setup()
 
     UNITY_BEGIN();    // IMPORTANT LINE!
 
+    testInitialisation();
     testTables();
     testTable2d();
